pyproject.toml
[project]
name = "medical-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "neo4j-graphrag>=1.8.0",
    "neo4j-rust-ext>=5.28.1.0",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
